refactor: relocate explore page from (main) to (extras) directory
